- Namespace reshuffling, making exports a lot easier for downstream users
