+++ src/index.ts
@@ -1,1 +1,2 @@
-
+import * as Trx from './coin/trx';
+export { Trx };
